# feat: add support for static cache config