--- a/drydock/patches/kustomization-resources
+++ b/drydock/patches/kustomization-resources
@@ -12,6 +12,7 @@
 - plugins/drydock/k8s/ingress/cms.yml
 - plugins/drydock/k8s/ingress/mfe.yml
 - plugins/drydock/k8s/ingress/extra-hosts.yml
+- plugins/drydock/k8s/ingress/static-cache.yml
 {%- endif %}
 {% if DRYDOCK_DEBUG -%}
 - plugins/drydock/k8s/debug/deployments.yml