--- a/drydock/plugin.py
+++ b/drydock/plugin.py
@@ -162,6 +162,7 @@ config = {
             "lms-worker",
             "cms-worker",
         ],
+        "NGINX_STATIC_CACHE_CONFIG": {},
     },
     # Add here settings that don't have a reasonable default for all users. For
     # instance: passwords, secret keys, etc.