### Description

This PR allows the setup of a configurable static cache running on NGINX. Make sure to configure the following snippet to your nginx HTTP config:

```nginx
proxy_cache_path /tmp/nginx_cache levels=1:2 keys_zone=static-cache:32m use_temp_path=off max_size=4g inactive=24h;
proxy_cache_key $scheme$proxy_host$request_uri;
proxy_cache_lock on;
proxy_cache_use_stale updating;
proxy_ignore_headers Cache-Control;
```

### Benchmarks:

Using k6 for fetching multiple static files.

Before:
```
          /\      |‾‾| /‾‾/   /‾‾/   
     /\  /  \     |  |/  /   /  /    
    /  \/    \    |     (   /   ‾‾\  
   /          \   |  |\  \ |  (‾)  | 
  / __________ \  |__| \__\ \_____/ .io

     execution: local
        script: openedx-testing/mfe/request_static_files.js
        output: -

     scenarios: (100.00%) 1 scenario, 400 max VUs, 1m0s max duration (incl. graceful stop):
              * default: 400 looping VUs for 30s (gracefulStop: 30s)

     █ MFE static files

     data_received........: 413 MB 12 MB/s
     group_duration.......: avg=4.46s    min=2.07s    med=4.12s    max=9.52s p(90)=6.37s p(95)=7.88s
     http_req_duration....: avg=2.11s    min=192.17ms med=1.99s    max=6.37s p(90)=3.38s p(95)=3.95s
     http_req_failed......: 0.00%  ✓ 0          ✗ 4704 
     http_req_receiving...: avg=696.22ms min=71.23µs  med=154.75ms max=3.97s p(90)=1.85s p(95)=2.26s
     http_req_waiting.....: avg=1.41s    min=158.02ms med=1.31s    max=4.88s p(90)=2.31s p(95)=2.73s
     http_reqs............: 4704   132.536314/s
     vus..................: 3      min=3        max=400
```
After:
```
          /\      |‾‾| /‾‾/   /‾‾/   
     /\  /  \     |  |/  /   /  /    
    /  \/    \    |     (   /   ‾‾\  
   /          \   |  |\  \ |  (‾)  | 
  / __________ \  |__| \__\ \_____/ .io

     execution: local
        script: openedx-testing/mfe/request_static_files.js
        output: -

     scenarios: (100.00%) 1 scenario, 400 max VUs, 1m0s max duration (incl. graceful stop):
              * default: 400 looping VUs for 30s (gracefulStop: 30s)

     █ MFE static files

     data_received........: 768 MB 22 MB/s
     group_duration.......: avg=1.88s    min=285.36ms med=1.45s    max=12.83s p(90)=4.22s    p(95)=5.14s   
     http_req_duration....: avg=864.04ms min=133.68ms med=333.14ms max=9.8s   p(90)=2.2s     p(95)=3s      
     http_req_failed......: 0.00%  ✓ 0          ✗ 8756 
     http_req_receiving...: avg=617.4ms  min=94.65µs  med=5.2ms    max=8.47s  p(90)=1.92s    p(95)=2.65s   
     http_req_waiting.....: avg=246.54ms min=132.95ms med=154.04ms max=3.15s  p(90)=510.83ms p(95)=768.91ms
     http_reqs............: 8756   253.325541/s
     vus..................: 7      min=7        max=400
```